separating lib build from wheel build
