+++ cpp/deps/requirements.txt
@@ -1,0 +1,7 @@
+wheel>=0.37.1
+nanobind>=2.1.0
+scikit-build-core>=0.10
+auditwheel>=6.1.0
+patchelf>=0.17.2
+toml>=0.10.2
+
