+++ cpp/docker/Dockerfile.glibc.wheeler
@@ -1,24 +1,10 @@
 # See https://github.com/pypa/manylinux for more information
 FROM quay.io/pypa/manylinux_2_28_x86_64
 
-ENV MBEDTLS_VERSION=3.6.2
 ENV PYTHON_PATH=/opt/python/cp310-cp310
 ENV PYTHON_EXECUTABLE=${PYTHON_PATH}/bin/python3.10
 
 COPY ./deps /tmp/deps
-
-# Install `mbedtls` from source
-# (there was no pre-built package available in the AlmaLinux repos)
-RUN cd /tmp \
-    && tar jxf ./deps/mbedtls-${MBEDTLS_VERSION}.tar.bz2 \
-    && cd mbedtls-${MBEDTLS_VERSION} \
-    && mkdir build \
-    && cd build \
-    && export CFLAGS="$CFLAGS -fPIC -Ofast" \
-    && cmake .. \
-    && make -j$(nproc) \
-    && make install \
-    && rm -rf /tmp/mbedtls-${MBEDTLS_VERSION}
 
 RUN ${PYTHON_EXECUTABLE} -m pip install -r /tmp/deps/requirements.txt
