+++ cpp/docker/Dockerfile.packer
@@ -1,0 +1,44 @@
+FROM python:3.10-slim-bookworm
+
+ENV MBEDTLS_VERSION=3.6.2
+ENV CMAKE_VERSION=3.30.5
+
+RUN apt-get update \
+    && apt-get install -y \
+    build-essential \
+    libssl-dev \
+    tar \
+    bzip2 \
+    file \
+    rpm \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/*
+
+COPY ./deps /tmp/deps
+
+# Install `cmake` from source
+RUN cd /tmp \
+    && tar zxf ./deps/cmake-${CMAKE_VERSION}.tar.gz \
+    && cd cmake-${CMAKE_VERSION} \
+    && ./bootstrap --parallel=$(nproc) \
+    && make -j$(nproc) \
+    && make install \
+    && rm -rf /tmp/cmake-${CMAKE_VERSION}
+
+# Install `mbedtls` from source
+RUN cd /tmp \
+    && tar jxf ./deps/mbedtls-${MBEDTLS_VERSION}.tar.bz2 \
+    && cd mbedtls-${MBEDTLS_VERSION} \
+    && mkdir build \
+    && cd build \
+    && export CFLAGS="$CFLAGS -fPIC -Ofast" \
+    && cmake .. \
+    && make -j$(nproc) \
+    && make install \
+    && rm -rf /tmp/mbedtls-${MBEDTLS_VERSION}
+
+RUN rm -rf /tmp/deps
+
+WORKDIR /hyperon_das_atomdb_cpp
+
+CMD ["bash", "-x", "./scripts/_build_packs.sh"]
