+++ cpp/lib/cmake/Packing.cmake
@@ -1,3 +1,4 @@
+set(CPACK_DEBIAN_PACKAGE_SHLIBDEPS ON)
 set(CPACK_PACKAGE_NAME ${PROJECT_NAME} CACHE STRING "hyperon_das_atomdb_cpp")
 set(
   CPACK_PACKAGE_DESCRIPTION_SUMMARY "Atom Space DB for Hyperon DAS"
