+++ cpp/lib/include/expression_hasher.h
@@ -23,8 +23,6 @@
  */
 #pragma once
 
-#include <mbedtls/md5.h>
-
 #include <stdexcept>
 #include <string>
 
@@ -60,23 +58,7 @@ class ExpressionHasher {
      * @param input The input string to be hashed.
      * @return A string representing the MD5 hash of the input.
      */
-    static const string compute_hash(const string& input) {
-        auto ctx = unique_ptr<mbedtls_md5_context, decltype(&mbedtls_md5_free)>(
-            new mbedtls_md5_context(), &mbedtls_md5_free);
-        mbedtls_md5_init(ctx.get());
-        uchar md5_buffer[MD5_BUFFER_SIZE];
-        if (mbedtls_md5_starts(ctx.get()) != 0 or
-            mbedtls_md5_update(ctx.get(), (const uchar*) input.c_str(), input.length()) != 0 or
-            mbedtls_md5_finish(ctx.get(), md5_buffer) != 0) {
-            throw runtime_error("Failed to compute MD5 hash");
-        }
-        char hash[2 * MD5_BUFFER_SIZE + 1];
-        for (unsigned int i = 0; i < MD5_BUFFER_SIZE; i++) {
-            sprintf(hash + 2 * i, "%02x", md5_buffer[i]);
-        }
-        hash[2 * MD5_BUFFER_SIZE] = '\0';
-        return move(string(hash));
-    }
+    static const string compute_hash(const string& input);
 
     /**
      * @brief Generates a hash for a named type.
@@ -84,7 +66,7 @@ class ExpressionHasher {
      * @param name The name of the type.
      * @return A string representing the hash of the named type.
      */
-    static const string named_type_hash(const string& name) { return compute_hash(name); }
+    static const string named_type_hash(const string& name);
 
     /**
      * @brief Generates a hash for a terminal expression.
@@ -94,13 +76,7 @@ class ExpressionHasher {
      * @return A string representing the hash of the terminal expression.
      * @throws invalid_argument if the terminal name is too large.
      */
-    static const string terminal_hash(const string& type, const string& name) {
-        if (type.length() + name.length() >= MAX_HASHABLE_STRING_SIZE) {
-            throw invalid_argument("Invalid (too large) terminal name");
-        }
-        string hashable_string = type + JOINING_CHAR + name;
-        return move(compute_hash(hashable_string));
-    }
+    static const string terminal_hash(const string& type, const string& name);
 
     /**
      * @brief Generates a hash for a composite expression.
@@ -108,21 +84,7 @@ class ExpressionHasher {
      * @param elements A vector of strings representing the elements of the composite expression.
      * @return A string representing the hash of the composite expression.
      */
-    static const string composite_hash(const StringList& elements) {
-        if (elements.size() == 1) {
-            return elements[0];
-        }
-
-        string hashable_string;
-        if (not elements.empty()) {
-            for (const auto& element : elements) {
-                hashable_string += element + JOINING_CHAR;
-            }
-            hashable_string.pop_back();  // remove the last joining character
-        }
-
-        return move(compute_hash(hashable_string));
-    }
+    static const string composite_hash(const StringList& elements);
 
     /**
      * @brief Generates a composite hash from a base hash.
@@ -133,7 +95,7 @@ class ExpressionHasher {
      * @param hash_base A string representing the base hash.
      * @return A string representing the composite hash generated from the base hash.
      */
-    static const string composite_hash(const string& hash_base) { return hash_base; }
+    static const string composite_hash(const string& hash_base);
 
     /**
      * @brief Generates a hash for an expression.
@@ -142,11 +104,7 @@ class ExpressionHasher {
      * @param elements A vector of strings representing the elements of the expression.
      * @return A string representing the hash of the expression.
      */
-    static const string expression_hash(const string& type_hash, const StringList& elements) {
-        StringList composite({type_hash});
-        composite.insert(composite.end(), elements.begin(), elements.end());
-        return move(composite_hash(composite));
-    }
+    static const string expression_hash(const string& type_hash, const StringList& elements);
 };
 
 }  // namespace atomdb
