+++ cpp/lib/scripts/_create_static_library.sh
@@ -1,0 +1,3 @@
+#!/bin/bash -x
+
+ar qc ${@}
