--- cpp/nanobind/deps/requirements.txt
@@ -1,6 +1,0 @@
-wheel>=0.37.1
-nanobind>=2.1.0
-scikit-build-core>=0.10
-auditwheel>=6.1.0
-patchelf>=0.17.2
-
