+++ cpp/nanobind/include/bind_helpers.h
@@ -14,6 +14,7 @@
  */
 #pragma once
 
+#include <hyperon_das_atomdb_cpp/database.h>
 #include <hyperon_das_atomdb_cpp/document_types.h>
 #include <hyperon_das_atomdb_cpp/type_aliases.h>
 #include <nanobind/nanobind.h>
