+++ cpp/scripts/build_packer_docker_image.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+source $(dirname $0)/common.sh
+
+docker buildx build -t ${PACKER_CONTAINER_NAME} \
+  --load -f ${PACKER_DOCKER_FILE} .
+
